requirements.txt
gmpy2==2.0.8
pycryptodome==3.12.0
